fixed build issues related to errors in patch scripts

+++ build/patch-legacy-libs.sh
@@ -9,14 +9,16 @@ mvn dependency:get -DgroupId=org.apache.commons -DartifactId=commons-text -Dvers
 
 COMMONS_LANG_DIR=$M2_PATH/org/apache/commons/commons-lang3/3.12.0
 jdeps --generate-module-info . $COMMONS_LANG_DIR/commons-lang3-3.12.0.jar
-javac --patch-module org.apache.commons.lang3=$COMMONS_LANG_DIR/commons-lang3-3.12.0.jar org.apache.commons.lang3/module-info.java
-jar uf $COMMONS_LANG_DIR/commons-lang3-3.12.0.jar -C org.apache.commons.lang3 module-info.class
-
-rm -r org.apache.commons.lang3
+if [ $? -eq 0 ]; then
+	javac --patch-module org.apache.commons.lang3=$COMMONS_LANG_DIR/commons-lang3-3.12.0.jar org.apache.commons.lang3/module-info.java
+	jar uf $COMMONS_LANG_DIR/commons-lang3-3.12.0.jar -C org.apache.commons.lang3 module-info.class
+	rm -r org.apache.commons.lang3
+fi
 
 COMMONS_TEXT_DIR=$M2_PATH/org/apache/commons/commons-text/1.10.0
 jdeps --module-path $COMMONS_LANG_DIR/commons-lang3-3.12.0.jar --generate-module-info . $COMMONS_TEXT_DIR/commons-text-1.10.0.jar
-javac --module-path $COMMONS_LANG_DIR/commons-lang3-3.12.0.jar --patch-module org.apache.commons.text=$COMMONS_TEXT_DIR/commons-text-1.10.0.jar org.apache.commons.text/module-info.java
-jar uf $COMMONS_TEXT_DIR/commons-text-1.10.0.jar -C org.apache.commons.text module-info.class
-
-rm -r org.apache.commons.text
+if [ $? -eq 0 ]; then
+	javac --module-path $COMMONS_LANG_DIR/commons-lang3-3.12.0.jar --patch-module org.apache.commons.text=$COMMONS_TEXT_DIR/commons-text-1.10.0.jar org.apache.commons.text/module-info.java
+	jar uf $COMMONS_TEXT_DIR/commons-text-1.10.0.jar -C org.apache.commons.text module-info.class
+	rm -r org.apache.commons.text
+fi
